Add some tests
