+++ app/tests/pages/ciip-application.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import CiipApplication from '../../pages/ciip-application';
+import CiipApplication from '../../containers/pageContainers/CiipApplication';
 
 const query = {
   allFormJsons: {
@@ -12,14 +12,6 @@ const query = {
 it('It matches the last accepted Snapshot', () => {
   const wrapper = shallow(<CiipApplication />);
   expect(wrapper).toMatchSnapshot();
-});
-
-// It renders a welcome message
-
-it('It renders a welcome message', () => {
-  const wrapper = shallow(<CiipApplication />);
-  const text = wrapper.find('p#welcome-message').text();
-  expect(text).toContain('Welcome');
 });
 
 it('It passes a query to the ApplicationWizard component', () => {
